核心升级到了1.3.0。
# 1.3.0 更新日志
1. 重写了文件传输模块，增强了安全性。
2. 支持以图片的形式发送消息。

+++ src/Core/src/protocol/helper/fileTransfer/FileReceiveTask.java
@@ -3,6 +3,11 @@ package protocol.helper.fileTransfer;
 import mutil.file.FileOccupiedException;
 import mutil.file.WriteOnlyFile;
 import mutil.uuidLocator.IUuidLocatable;
+import mutil.uuidLocator.UuidConflictException;
+import protocol.dataPack.DataPack;
+import protocol.dataPack.FileTransferType;
+import protocol.dataPack.UploadResultPack;
+import protocol.helper.data.PackageTooLargeException;
 
 import java.io.IOException;
 import java.util.UUID;
@@ -10,44 +15,85 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 public abstract class FileReceiveTask implements IUuidLocatable {
-    protected final UUID uuid;
-    protected final long fileSize;
+    protected UUID senderTaskId, receiverTaskId;
     private long transferredSize = 0;
 
-    protected final WriteOnlyFile file;
+    protected WriteOnlyFile file;
+    protected long fileSize;
+    protected UUID localFileId;
 
-    protected abstract void onEndSuccess();
-    protected abstract void onEndFail();
+    protected abstract FileTransferType getFileTransferType();
+    protected abstract void send(DataPack dataPack) throws IOException, PackageTooLargeException;
 
-    public FileReceiveTask(UUID uuid,WriteOnlyFile file,long fileSize){
-        this.uuid = uuid;
-        this.file = file;
-        this.fileSize = fileSize;
+    protected void onEndSucceed(){
+        if(senderTaskId==null) return;
+        try{
+            send(new UploadResultPack(
+                    senderTaskId,
+                    file.getFileObject().getFileId(),
+                    true,
+                    ""
+            ));
+        }catch (IOException ignored){
+        }catch (PackageTooLargeException e){
+            throw new RuntimeException(e);
+        }
     }
 
-    public void start(){
-        onCreate();
+    protected void onEndFailed(String reason){
+        if(senderTaskId==null) return;
+        try{
+            send(new UploadResultPack(
+                    senderTaskId,
+                    new UUID(0,0),
+                    false,
+                    reason
+            ));
+        }catch (IOException ignored){
+        }catch (PackageTooLargeException e){
+            throw new RuntimeException(e);
+        }
+    }
+
+    public void init(){
+        while(true){
+            try{
+                receiverTaskId = UUID.randomUUID();
+                onCreate();
+                break;
+            }catch (UuidConflictException ignored){
+            }
+        }
     }
 
     protected void showInfo(String info){}
     protected void setTransferProgress(double progress){}
 
     public void end(){
-        file.close();
-        if(file.length()==this.fileSize){
-            this.onEndSuccess();
-        }else{
-            this.onEndFail();
-            try{
-                file.getFileObject().delete();
-            }catch (FileOccupiedException e){
-                Logger.getGlobal().log(Level.WARNING,"File delete failed.",e);
+        if(file!=null){
+            file.close();
+            if(file.length()==this.fileSize){
+                this.onEndSucceed();
+            }else{
+                this.onEndFailed("Actual received size does not match the request.");
+                try{
+                    file.getFileObject().delete();
+                }catch (FileOccupiedException e){
+                    Logger.getLogger("IMCore").log(
+                            Level.WARNING,
+                            String.format("File %s is occupied and cannot be deleted.",file.getFileObject().getFile().getAbsoluteFile()),
+                            e
+                    );
+                }
             }
+        }else{
+            this.onEndFailed("Nothing received.");
         }
         onDelete();
     }
 
     public void end(String reason){
+        end();
         showInfo(reason);
     }
 
@@ -59,7 +105,28 @@ public abstract class FileReceiveTask implements IUuidLocatable {
 
     @Override
     public UUID getUuid() {
-        return uuid;
+        return receiverTaskId;
+    }
+
+    public UUID getReceiverTaskId() {
+        return receiverTaskId;
+    }
+
+    public void setWriteOnlyFile(WriteOnlyFile file) {
+        this.localFileId = file.getFileObject().getFileId();
+        this.file = file;
+    }
+
+    public void setFileSize(long size){
+        this.fileSize = size;
+    }
+
+    public void setSenderTaskId(UUID senderTaskId) {
+        this.senderTaskId = senderTaskId;
+    }
+
+    public UUID getLocalFileId() {
+        return localFileId;
     }
 
 }
